+++ Model/__init__.py
@@ -1,0 +1,4 @@
+from .chain import Chain
+from .model import LLM
+
+__all__ = ['Chain', 'LLM']
