+++ main.py
@@ -1,0 +1,4 @@
+from Model import *
+
+if __name__ == "__main__":
+    Chain.test_chain()
